pyproject.toml
[project]
name = "minauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "pyjwt>=2.10.1",
]
